chore(examples): add SDK v2 and SDK v3 examples
