+++ examples/shared-base/entities/organisation.entity.ts
@@ -1,0 +1,23 @@
+import {
+  Attribute,
+  AutoGenerateAttribute,
+  AUTO_GENERATE_ATTRIBUTE_STRATEGY,
+  Entity,
+} from '@typedorm/common';
+
+@Entity({
+  name: 'organisation',
+  primaryKey: {
+    partitionKey: 'ORG#{{id}}',
+    sortKey: 'ORG#{{id}}',
+  },
+})
+export class Organisation {
+  @AutoGenerateAttribute({
+    strategy: AUTO_GENERATE_ATTRIBUTE_STRATEGY.UUID4,
+  })
+  id: string;
+
+  @Attribute()
+  name: string;
+}
