+++ examples/shared-base/index.ts
@@ -1,0 +1,3 @@
+import 'reflect-metadata';
+export * from './table';
+export * from './entities/organisation.entity';
