+++ examples/shared-base/rollup.config.js
