+++ examples/shared-base/table.ts
@@ -1,0 +1,7 @@
+import {Table} from '@typedorm/common';
+
+export const myGlobalTable = new Table({
+  name: 'test-table',
+  partitionKey: 'PK',
+  sortKey: 'SK',
+});
